pyproject.toml
[project]
name = "pomni"
version = "0.1.0"
description = "JAX Code assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "patdb>=0.1.0",
    "pylsp-rope>=0.1.17",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.2",
    "pdbpp>=0.11.6",
    "python-lsp-server>=1.12.2",
    "ruff>=0.11.12",
]
